requirements.txt
streamlit
pandas
numpy
python-dotenv
requests
openai
google-generativeai
groq
SQLAlchemy
psycopg2-binary
passlib
bcrypt
PyJWT
fpdf2
Pillow
pypdf2
xlsxwriter
python-docx
openpyxl
arabic-reshaper
python-bidi
plotly
feedparser
beautifulsoup4
folium
streamlit-folium
ezdxf
reportlab
svglib
pydantic
# Updated 2026-03-24 - Removed pygments/rich/markdown-it-py (streamlit transitive deps - caused conflict)
